mondaynet magic

+++ mondaynet/mondaynet-start.sh
@@ -26,6 +26,14 @@ fi
 #
 builddir=$buildroot/tezos
 
+if [ -f "$HOME/.cleanup" ]; then
+	echo "===> Cleaning up"
+	rm -rf tezos rustup-init.sh monday.txt logs fetch-params.sh \
+		.zcash-params .opam .rustup .cargo .cache $HOME/.cleanup \
+		$HOME/.firstrun "$HOME/.skipbuild"
+fi
+
+
 # Check dependencies
 #
 if [ ! -d $HOME/.zcash-params ]; then
@@ -35,11 +43,15 @@ if [ ! -d $HOME/.zcash-params ]; then
 	sh ./fetch-params.sh
 fi
 
+if [ ! -d $HOME/.cargo ]; then
+	rm -f "$HOME/.firstrun"
+fi
+
 # Check first run
 #
 if [ ! -f "$HOME/.firstrun" ]; then
-	sudo apt update
-	sudo apt install -y rsync git m4 build-essential patch unzip wget pkg-config libgmp-dev libev-dev libhidapi-dev libffi-dev opam jq zlib1g-dev bc autoconf
+	sudo apt-get update
+	sudo apt-get install -y rsync git m4 build-essential patch unzip wget pkg-config libgmp-dev libev-dev libhidapi-dev libffi-dev opam jq zlib1g-dev bc autoconf
 
 	echo "==> Installing rust"
 	echo ""
@@ -54,6 +66,7 @@ fi
 # Update the software to latest master branch of Octez
 #
 if [ ! -d $builddir ]; then
+	rm -f "$HOME/.skipbuild"
 	echo "===> Setting up software"
 	mkdir -p "$buildroot"
 	cd $buildroot
@@ -62,32 +75,47 @@ if [ ! -d $builddir ]; then
 	opam init --bare --yes
 fi
 
-eval $(opam env) 
 
-echo "===> Updating the branch"
-cd $builddir
-git checkout $branch
-git pull
-rm -rf _build _opam
+if [ ! -f "$HOME/.skipbuild" ]; then 
+	eval $(opam env) 
 
-echo "===> Rebuilding the software"
-make build-deps
-eval $(opam env) 
-make
+	echo "===> Updating the branch"
+	cd $builddir
+	git checkout $branch
+	git pull
+	rm -rf _build _opam
 
-if [ ! -f tezos-node ]; then
-	echo "XXX Build unsuccessful!"
+	echo "===> Rebuilding the software"
+	make build-deps
+	if [ $? != "0" ]; then
+		echo "XXX Failed to build dependencies"
+		exit 1
+	fi
+	eval $(opam env) 
+	make
+	if [ $? != "0" ]; then
+		echo "XXX Failed to build tezos"
+		exit 1
+	fi
+fi
+rm -f "$HOME/.skipbuild"
+
+if [ ! -f "$builddir/tezos-node" ]; then
+	echo "XXX No node binary!"
 	echo "EXITING"
 	exit 1
 fi
 
-# Reset baking account
-#
-echo "===> Resetting node and baking"
-echo ""
-rm -rf "$HOME/.tezos-node"
-# XXX also pieces in tezos client
-sleep 5
+if [ ! -f "$HOME/.skipreset" ]; then
+	# Reset baking account
+	#
+	echo "===> Resetting node and baking"
+	echo ""
+	rm -rf "$HOME/.tezos-node"
+	# XXX also pieces in tezos client
+	sleep 5
+fi
+rm -f "$HOME/.skipreset"
 
 # Start the node
 #
